workspaces/rabbit-steering/pyproject.toml
[project]
name = "rabbit-steering"
version = "0.0.1"
requires-python = ">=3.12,<3.13"
dependencies = [
    "nats-py",
    "adafruit-circuitpython-servokit",
    "adafruit-blinka",
    "Jetson.GPIO",
]
